Split USBDeviceShim into two
USBDeviceShim handles Yosemite and older
USBHostDeviceShim handles El Capitan and newer
